Iteration 6 - Introduce additional package for command classes
