+++ src/main/java/com/acme/dbo/txlog/Facade.java
@@ -1,5 +1,7 @@
 package com.acme.dbo.txlog;
 
+import com.acme.dbo.txlog.command.*;
+
 public class Facade {
     private static Controller controller = new Controller();
