+++ src/main/java/com/acme/dbo/txlog/command/Command.java
@@ -1,6 +1,8 @@
-package com.acme.dbo.txlog;
+package com.acme.dbo.txlog.command;
 
-interface Command {
+import com.acme.dbo.txlog.Controller;
+
+public interface Command {
 
     String getDecoratedState(int duplicationNum);
